% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_monitor.R
\name{pa_2utm}
\alias{pa_2utm}
\title{Reproject a sf object to UTM coordinates}
\usage{
pa_2utm(df, verbose = FALSE)
}
\arguments{
\item{df}{sf object to be reprojected to UTM coordinates}

\item{verbose}{whether to print operation details}
}
\value{
a sf object
}
\description{
Reproject a sf object to UTM coordinates
}
\details{
This function will attempt to automatically
  determine the adequate UTM zone and reproject a sf
  object,
}
\examples{
\donttest{
## for examples, see vignette pacu
}

}
\author{
Caio dos Santos and Fernando Miguez
}
