% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\alias{merge.trial}
\title{Merge trial objects}
\usage{
\method{merge}{trial}(...)
}
\arguments{
\item{...}{\code{trial} objects to merge.}
}
\value{
An object of class \code{trial}.
}
\description{
Combine two or more \code{trial} objects into a single
  \code{trial} object. This is an S3 method for \code{base::merge} and
  is useful when a field experiment spans multiple processing runs that
  need to be joined before analysis.
}
