% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\alias{merge.trial}
\title{Merge trial objects}
\usage{
\method{merge}{trial}(...)
}
\arguments{
\item{...}{pa_trial objects}
}
\value{
object of class "trial"
}
\description{
Generic merge functionalities for pacu objects
}
