% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.yield}
\alias{print.dslist}
\title{Print a pacu object}
\usage{
\method{print}{yield}(x, ...)

\method{print}{dslist}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{additional arguments. None used currently.}
}
\description{
These functions print meaningful information from
pacu objects.
}
