% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacu.R
\docType{data}
\name{pacu.options}
\alias{pacu.options}
\title{Environment which stores PACU options}
\format{
An object of class \code{environment} of length 3.
}
\usage{
pacu.options
}
\value{
This is an environment, not a function, so nothing is returned.
}
\description{
Environment which can store the path to the executable, warning settings and
             where examples are located.
             Creating an environment avoids the use of global variables or other similar practices
             which would have possible undesirable consequences.
}
\details{
Environment which stores PACU options
}
\examples{
\dontrun{
names(pacu.options)
pacu_options(exe.path = "some-new-path-to-executable")
pacu.options$exe.path
}

}
\keyword{datasets}
