% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacu.R
\name{pacu_options}
\alias{pacu_options}
\title{Setting some options for the package}
\usage{
pacu_options(exe.path = NA)
}
\arguments{
\item{exe.path}{path to apsim executable. White spaces are not allowed.}
}
\value{
as a side effect it modifies the \sQuote{pacu.options} environment.
}
\description{
Set the path to the APSIM-X executable, examples and warning suppression.
}
\details{
Set pacu options
}
\examples{
\dontrun{
names(pacu.options)
pacu_options(exe.path = "some-new-path-to-executable")
pacu.options$exe.path
}
}
