% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacu.R
\name{pacu_options}
\alias{pacu_options}
\title{Setting some options for the package}
\usage{
pacu_options(suppress.warnings = FALSE, suppress.messages = FALSE)
}
\arguments{
\item{suppress.warnings}{whether to suppress warning messages}

\item{suppress.messages}{whether to suppress  messages}
}
\value{
as a side effect it modifies the \sQuote{pacu.options} environment.
}
\description{
Set settings regarding messages and default behaviors of the package
}
\details{
Set pacu options
}
\examples{
\dontrun{
names(pacu.options)
pacu_options(exe.path = "some-new-path-to-executable")
pacu.options$exe.path
}
}
